requirement.txt
# Web Framework
flask==3.0.0
numpy
scikit-learn

# Database and ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Environment Variables
python-dotenv==1.0.0

# Serialization
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Additional Utilitiescl
flask-cors==4.0.0
requests==2.31.0

# Development and Testing
pytest==7.4.4